# api/src/specs/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"GameTypes": {
				"enum": [
					"free",
					"tarot",
					"poker"
				],
				"type": "string"
			},
			"Game": {
				"properties": {
					"name": {
						"type": "string"
					},
					"game_type": {
						"$ref": "#/components/schemas/GameTypes"
					},
					"start_time": {
						"type": "string"
					},
					"end_time": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"game_type",
					"start_time",
					"end_time"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Games": {
				"items": {
					"$ref": "#/components/schemas/Game"
				},
				"type": "array"
			},
			"Player": {
				"properties": {
					"game_id": {
						"type": "number",
						"format": "double"
					},
					"player_name": {
						"type": "string"
					},
					"player_token": {
						"type": "string"
					},
					"creation_time": {
						"type": "string"
					}
				},
				"required": [
					"game_id",
					"player_name",
					"player_token",
					"creation_time"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Players": {
				"items": {
					"$ref": "#/components/schemas/Player"
				},
				"type": "array"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"description": "api for website platform for games",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Pierre-Marie Airiau"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/games/all": {
			"get": {
				"operationId": "GetGames",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Games"
								}
							}
						}
					}
				},
				"description": "get all games registered",
				"security": [],
				"parameters": []
			}
		},
		"/games/get/{id}": {
			"get": {
				"operationId": "GetGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Game"
								}
							}
						}
					}
				},
				"description": "return infos about the Game Id requested",
				"security": [],
				"parameters": [
					{
						"description": "gameId",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/games/new-game/{gameType}": {
			"post": {
				"operationId": "CreateGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"description": "Create a new game",
				"security": [],
				"parameters": [
					{
						"description": "which type of game do you want to create",
						"in": "path",
						"name": "gameType",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/GameTypes"
						}
					}
				]
			}
		},
		"/health": {
			"get": {
				"operationId": "GetHealth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "get all games registered",
				"security": [],
				"parameters": []
			}
		},
		"/players/get/{gameId}": {
			"get": {
				"operationId": "GetPlayers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Players"
								}
							}
						}
					}
				},
				"description": "get player infos for a specific game",
				"security": [],
				"parameters": [
					{
						"description": "gameId for which you want players",
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/players/new-player": {
			"post": {
				"operationId": "CreatePlayer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "Register a new player for a game",
				"security": [],
				"parameters": [
					{
						"description": "name of the player you want to register",
						"in": "query",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "gameId for which you want to register the player",
						"in": "query",
						"name": "gameId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}